requirements.txt
discord.py
requests
python-dotenv
bybit-ws